requirements.txt
scikit-learn
apache-airflow-providers-microsoft-mssql[common.sql]
pymssql